requirements.txt
Django>=4.0
djangorestframework>=3.12
psycopg2-binary>=2.9
